\name{bread}
\alias{bread}
\alias{bread.ppml}


\title{
Bread for Sandwiches
}

\description{
This method is based on the function \code{bread} from the \sQuote{sandwich} package for the object of class \code{ppml}.
}

\usage{
\S3method{bread}{ppml}(object, ...)
}

\arguments{
  \item{object}{An object of class \code{ppml}.}
  \item{...}{Currently not used.}
}

\value{
Returns the bread matrix.
}



\references{
  Zeileis A (2006), Object-Oriented Computation of Sandwich Estimators.
  \emph{Journal of Statistical Software}, \bold{16}(9), 1--16.
  URL \url{http://www.jstatsoft.org/v16/i09/}.
}

\author{
Lukas Kleinheinz
}


\concept{bread}

